added toonshading, spotlight, started textures
toon shading and spotlight werent working because they were not properly initialised, this is now fixed.

spotlight is mapped to the bottom of the rollercoaster, should be mapped to the front like a torch! mess with the light direction a bit.

started adding textures for the models, however not sure how to add for only one model, and skybox textures are now flipped.